+++ AccountingSystem/Views/Transactions/ReceiptsIssued/frmReceiptsEdit.cs
@@ -61,8 +61,8 @@ namespace AccountingSystem.Views.Transactions.ReceiptsIssued
                 var riModel = new ReceiptsIssuedModel()
                 {
                     Id = uc.Id,
-                    CoId = Convert.ToInt16(uc.cmbcollector.SelectedValue),
-                    RId = Convert.ToInt16(uc.cmbreceipt.SelectedValue),
+                    CoId = Convert.ToInt32(uc.cmbcollector.SelectedValue),
+                    RId = Convert.ToInt32(uc.cmbreceipt.SelectedValue),
                     Issued = uc.dtpissued.Value,
                     IssuedFrom = Convert.ToInt32(uc.txtfrom.Text.Trim()),
                     IssuedTo = Convert.ToInt32(uc.txtto.Text.Trim()),
@@ -70,7 +70,7 @@ namespace AccountingSystem.Views.Transactions.ReceiptsIssued
                 };
 
                 var riRepository = Factory.ReceiptsIssuedRepository();
-                if (Convert.ToInt16(uc.txtfrom.Text.Trim()) > Convert.ToInt16(uc.txtto.Text.Trim()))
+                if (Convert.ToInt32(uc.txtfrom.Text.Trim()) > Convert.ToInt32(uc.txtto.Text.Trim()))
                 {
                     Helper.MessageBoxError("Invalid Receipt!");
                     return false;
